Move tooltip relative to cursor at lower right
When the cursor was too far towards the bottom of the plots, the tooltip
extended below the window and changed window height, which changed the
cursor position if the window was scrolled all the way to the bottom.
This made the tooltip flicker funny.

When the cursor was too far towards the right of the plots, it extended
outside the edge of the app and was cut off.